replacing get.env to os.environ
Benefit of environ.get is that it access en variable dicitionary and hasnt need to check variable existence. It also validate tht variable exist in env before trying to access it. The point is that if variable doesnt exist in os.environ.get then return none which is better to check if variable is none before using it