Use Map function instead of list
by implementing map() in int() which in case applies to all elements of an iterablesplit function which returns a list of string. It then returns an iterator which is far more memory efficient than creting a new list with a list comphreension. Generally map() is better especially when you work with large data sets